+++ tests/primitives/test_country.py
@@ -2,13 +2,12 @@ import six
 from pytest import mark, raises
 
 from sqlalchemy_utils import Country, i18n
-from sqlalchemy_utils.primitives.currency import babel  # noqa
 
 
-@mark.skipif('babel is None')
+@mark.skipif('i18n.babel is None')
 class TestCountry(object):
     def setup_method(self, method):
-        i18n.get_locale = lambda: babel.Locale('en')
+        i18n.get_locale = lambda: i18n.babel.Locale('en')
 
     def test_init(self):
         assert Country(u'FI') == Country(Country(u'FI'))