+++ tests/primitives/test_currency.py
@@ -3,13 +3,12 @@ import six
 from pytest import mark, raises
 
 from sqlalchemy_utils import Currency, i18n
-from sqlalchemy_utils.primitives.currency import babel  # noqa
 
 
-@mark.skipif('babel is None')
+@mark.skipif('i18n.babel is None')
 class TestCurrency(object):
     def setup_method(self, method):
-        i18n.get_locale = lambda: babel.Locale('en')
+        i18n.get_locale = lambda: i18n.babel.Locale('en')
 
     def test_init(self):
         assert Currency('USD') == Currency(Currency('USD'))